hierCluster loads data from new route termdb/cluster
